split program up into multiple files
